+++ frontend/src/pages/IndexPage.jsx
@@ -16,6 +16,9 @@ export default function IndexPage( handleLogout ) {
     price: false
   })
 
+    console.log(typeof handleLogout)
+
+
   const user = useContext(UserContext)
   const host = 'http://localhost:2003'
   const dropdownIcon = <FontAwesomeIcon icon={faAngleDown} />
@@ -80,6 +83,11 @@ export default function IndexPage( handleLogout ) {
     <div className='flex flex-col min-h-screen'>
       <Header handleLogout={handleLogout} textColor='white'/>
       <Hero />
+
+      <div> 
+        {user && <h2 className='h2 text-center mt-16'>Welcome, {user}!</h2>}
+      </div>
+
       <div className='max-w-[1240px] grid grid-flow-col gap-12 justify-center mx-auto my-20'>
         <div className='h-80 w-64 rounded-md bg-[url("src/assets/gradient1.png")] bg-cover bg-bottom 
         flex justify-center items-center shadow-lg'>
@@ -94,68 +102,11 @@ export default function IndexPage( handleLogout ) {
           <a href='/'><h2 className='h2 text-center mx-8 uppercase'>Budget Sunscreens</h2></a>
         </div>
       </div>
-      <div>
-        <div> 
-          {user && <h2 className='h2 text-center mt-16'>Welcome, {user}!</h2>}
-        </div>
-        <div className='max-w-[1240px] grid grid-flow-col justify-start mx-auto mt-12 mb-10 gap-4'>
-          <h3 className='h3 my-auto'>Filter by</h3>
-          <div ref={dropdownRefs.form}>
-            <button className='btn-dropdown flex justify-between items-center' onClick={() => toggleDropdown('form')}> 
-              Form {dropdownIcon}
-            </button>
-            {dropdownStates.form && (
-              <div className='dropdown'>
-                <a href='' className='dropdown-link'>Lotion</a>
-                <a href='' className='dropdown-link'>Spray</a>
-                <a href='' className='dropdown-link'>Stick</a>
-              </div>
-            )}
-          </div>
-          <div ref={dropdownRefs.type}>
-            <button className='btn-dropdown flex justify-between items-center' onClick={() => toggleDropdown('type')}> 
-              Type {dropdownIcon}
-            </button>
-            {dropdownStates.type && (
-              <div className='dropdown'>
-                <a href='' className='dropdown-link'>Chemical</a>
-                <a href='' className='dropdown-link'>Mineral</a>
-                <a href='' className='dropdown-link'>Hybrid</a>
-              </div>
-            )}
-          </div>
-          <div ref={dropdownRefs.spf}>
-            <button className='btn-dropdown flex justify-between items-center' onClick={() => toggleDropdown('spf')}> 
-              SPF {dropdownIcon}
-            </button>
-            {dropdownStates.spf && (
-              <div className='dropdown'>
-                <a href='' className='dropdown-link'>15+</a>
-                <a href='' className='dropdown-link'>30+</a>
-                <a href='' className='dropdown-link'>50+</a>
-                <a href='' className='dropdown-link'>70+</a>
-              </div>
-            )}
-          </div>
-          <div ref={dropdownRefs.price}>
-            <button className='btn-dropdown flex justify-between items-center' onClick={() => toggleDropdown('price')}> 
-              Price {dropdownIcon} 
-            </button>
-            {dropdownStates.price && (
-              <div className='dropdown'>
-                <a href='' className='dropdown-link'>$</a>
-                <a href='' className='dropdown-link'>$$</a>
-                <a href='' className='dropdown-link'>$$$</a>
-                <a href='' className='dropdown-link'>$$$$</a>
-                <a href='' className='dropdown-link'>$$$$$</a>
-              </div>
-            )}
-          </div>
-        </div>
 
-        {/* Pass in list to Sunscreens component, which loads individual Sunscreen components */}
-        <Sunscreens sunscreens={sunscreens} />
+      <div>
+        <Sunscreens />
       </div>
+
       <Footer />
     </div>
     </>
